chore: moving http matters in api folder + extract config
